vetApp/requirements.txt
black==21.11b1
Django==4.0.2
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
isort==5.10.1
psycopg2-binary==2.9.3